+++ datapilot/__init__.py
@@ -4,6 +4,6 @@ Datapilot
 An intelligent Python library for exploratory data analysis.
 """
 
-from .api.analyze import analyze
+from .api import analyze
 
 __all__ = ["analyze"]